job-app-assistant/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
httpx==0.26.0
PyMuPDF==1.24.1
PyYAML==6.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
